requirements.txt
websocket-client>=1.6.0
python-dotenv>=1.0.0
pywin32>=306
openpyxl>=3.1.0
mcp[cli]>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
